Make scalar and tensor share the casting policy function.
